+++ lib/ui/uikit/src/components/PlDropdownRef/index.ts
@@ -1,0 +1,1 @@
+export { default as PlDropdownRef } from './PlDropdownRef.vue';